requirements.txt
beautifulsoup4==4.12.3
langchain==0.2.6
langchain_chroma==0.1.2
langchain_community==0.2.6
langchain_core==0.2.11
langchain_openai==0.1.14
langchain_text_splitters==0.2.2
openai==1.35.10
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
slack_sdk==3.30.0
streamlit==1.32.0
typing_extensions==4.12.2
gpt4all==2.7.0
